Spell mouse over, spell bar implemented.

+++ source/core/item_manager.hpp
@@ -152,7 +152,7 @@ struct item_spell_level_type
     int              maximum;
     int              current;
     float            experience;
-    int              next;
+    float            next;
 };
 
 struct item_spell_data_type
